# apps/server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Booking API",
    "version": "0.0.0"
  },
  "components": {
    "schemas": {
      "Booking": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "guestName": {
            "type": "string"
          },
          "roomType": {
            "type": "string",
            "enum": ["single", "double", "suite"]
          },
          "checkIn": {
            "type": "string",
            "format": "date"
          },
          "checkOut": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "confirmed", "cancelled"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "guestName", "roomType", "checkIn", "checkOut", "status", "createdAt"],
        "additionalProperties": false
      },
      "CreateBookingInput": {
        "type": "object",
        "properties": {
          "guestName": {
            "type": "string"
          },
          "roomType": {
            "type": "string",
            "enum": ["single", "double", "suite"]
          },
          "checkIn": {
            "type": "string",
            "format": "date"
          },
          "checkOut": {
            "type": "string",
            "format": "date"
          }
        },
        "required": ["guestName", "roomType", "checkIn", "checkOut"],
        "additionalProperties": false
      }
    },
    "parameters": {}
  },
  "paths": {
    "/api/bookings": {
      "get": {
        "summary": "List all bookings",
        "responses": {
          "200": {
            "description": "A list of bookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Booking"
                      }
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a booking",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Booking"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload"
          }
        }
      }
    },
    "/api/bookings/{id}": {
      "get": {
        "summary": "Get a booking by id",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The booking",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Booking"
                }
              }
            }
          },
          "400": {
            "description": "Invalid identifier format"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  }
}